requirements.txt
waitress
sqlalchemy
gunicorn
flask
flask-login
flask-sqlalchemy
python-decouple
